Create dollarToPulse.ino
This example builds on the basic bill acceptor example and adds a pulse output based on the currency note that was inserted. 
The pulse output pin can be configured via outPin, the delay for pulse on and pulse off can bet configured through their variables too, and finally the number of pulses per bill value can also be configured.